+++ NetworkLib/src/UDP/Packet.hpp
@@ -2,7 +2,7 @@
 
 #include <UDP/Datagram.hpp>
 #include <UDP/ChannelHeader.hpp>
-#include <cstdint>
+#include <Types.hpp>
 #include <array>
 
 namespace Bousk
@@ -13,8 +13,8 @@ namespace Bousk
 		{
 			struct Packet
 			{
-				using Id = uint16_t;
-				enum class Type : uint8_t
+				using Id = uint16;
+				enum class Type : uint8
 				{
 					FullMessage,
 					FirstFragment,
@@ -24,27 +24,27 @@ namespace Bousk
 				struct Header
 				{
 					Id id;
-					uint16_t size{ 0 };
+					uint16 size{ 0 };
 					Type type;
 				};
-				static constexpr uint16_t PacketMaxSize = Datagram::DataMaxSize - ChannelHeader::Size;
-				static constexpr uint16_t HeaderSize = sizeof(Header);
-				static constexpr uint16_t DataMaxSize = PacketMaxSize - HeaderSize;
-				static constexpr size_t MaxPacketsPerMessage = 32;
-				static constexpr size_t MaxMessageSize = MaxPacketsPerMessage * DataMaxSize;
+				static constexpr uint16 PacketMaxSize = Datagram::DataMaxSize - ChannelHeader::Size;
+				static constexpr uint16 HeaderSize = sizeof(Header);
+				static constexpr uint16 DataMaxSize = PacketMaxSize - HeaderSize;
+				static constexpr uint8 MaxPacketsPerMessage = 32;
+				static constexpr uint16 MaxMessageSize = MaxPacketsPerMessage * DataMaxSize;
 
 				Header header;
-				std::array<uint8_t, DataMaxSize> dataBuffer;
+				std::array<uint8, DataMaxSize> dataBuffer;
 
 				inline Id id() const { return header.id; }
 				inline Type type() const { return header.type; }
-				inline uint8_t* data() { return dataBuffer.data(); }
-				inline const uint8_t* data() const { return dataBuffer.data(); }
-				inline uint16_t datasize() const { return header.size; }
+				inline uint8* data() { return dataBuffer.data(); }
+				inline const uint8* data() const { return dataBuffer.data(); }
+				inline uint16 datasize() const { return header.size; }
 				//!< Full buffer : header + data
-				inline const uint8_t* buffer() const { return reinterpret_cast<const uint8_t*>(this); }
+				inline const uint8* buffer() const { return reinterpret_cast<const uint8*>(this); }
 				//!< Packet full size : header + data
-				inline uint16_t size() const { return HeaderSize + header.size; }
+				inline uint16 size() const { return HeaderSize + header.size; }
 			};
 		}
 	}
