+++ NetworkLib/src/UDP/Protocols/ReliableOrdered.hpp
@@ -27,14 +27,14 @@ namespace Bousk
 					ReliableOrdered() = default;
 					~ReliableOrdered() override = default;
 
-					void queue(std::vector<uint8_t>&& msgData) override { mMultiplexer.queue(std::move(msgData)); }
-					size_t serialize(uint8_t* buffer, const size_t buffersize, const Datagram::ID datagramId) override { return mMultiplexer.serialize(buffer, buffersize, datagramId); }
+					void queue(std::vector<uint8>&& msgData) override { mMultiplexer.queue(std::move(msgData)); }
+					uint16 serialize(uint8* buffer, const uint16 buffersize, const Datagram::ID datagramId) override { return mMultiplexer.serialize(buffer, buffersize, datagramId); }
 
 					void onDatagramAcked(const Datagram::ID datagramId) override { mMultiplexer.onDatagramAcked(datagramId); }
 					void onDatagramLost(const Datagram::ID datagramId) override { mMultiplexer.onDatagramLost(datagramId); }
 
-					void onDataReceived(const uint8_t* data, const size_t datasize) override { mDemultiplexer.onDataReceived(data, datasize); }
-					std::vector<std::vector<uint8_t>> process() override { return mDemultiplexer.process(); }
+					void onDataReceived(const uint8* data, const uint16 datasize) override { mDemultiplexer.onDataReceived(data, datasize); }
+					std::vector<std::vector<uint8>> process() override { return mDemultiplexer.process(); }
 
 					virtual bool isReliable() const { return true; }
 
@@ -46,8 +46,8 @@ namespace Bousk
 						Multiplexer() = default;
 						~Multiplexer() = default;
 
-						void queue(std::vector<uint8_t>&& msgData);
-						size_t serialize(uint8_t* buffer, const size_t buffersize, const Datagram::ID datagramId);
+						void queue(std::vector<uint8>&& msgData);
+						uint16 serialize(uint8* buffer, const uint16 buffersize, const Datagram::ID datagramId);
 
 						void onDatagramAcked(const Datagram::ID datagramId);
 						void onDatagramLost(const Datagram::ID datagramId);
@@ -82,8 +82,8 @@ namespace Bousk
 						Demultiplexer() = default;
 						~Demultiplexer() = default;
 
-						void onDataReceived(const uint8_t* data, const size_t datasize);
-						std::vector<std::vector<uint8_t>> process();
+						void onDataReceived(const uint8* data, const uint16 datasize);
+						std::vector<std::vector<uint8>> process();
 
 					private:
 						void onPacketReceived(const Packet* pckt);
