+++ NetworkLib/src/UDP/Protocols/UnreliableOrdered.hpp
@@ -25,11 +25,11 @@ namespace Bousk
 					UnreliableOrdered() = default;
 					~UnreliableOrdered() override = default;
 
-					void queue(std::vector<uint8_t>&& msgData) override { mMultiplexer.queue(std::move(msgData)); }
-					size_t serialize(uint8_t* buffer, size_t buffersize, Datagram::ID datagramId) override { return mMultiplexer.serialize(buffer, buffersize, datagramId); }
+					void queue(std::vector<uint8>&& msgData) override { mMultiplexer.queue(std::move(msgData)); }
+					uint16 serialize(uint8* buffer, uint16 buffersize, Datagram::ID datagramId) override { return mMultiplexer.serialize(buffer, buffersize, datagramId); }
 
-					void onDataReceived(const uint8_t* data, size_t datasize) override { mDemultiplexer.onDataReceived(data, datasize); }
-					std::vector<std::vector<uint8_t>> process() override { return mDemultiplexer.process(); }
+					void onDataReceived(const uint8* data, uint16 datasize) override { mDemultiplexer.onDataReceived(data, datasize); }
+					std::vector<std::vector<uint8>> process() override { return mDemultiplexer.process(); }
 
 					virtual bool isReliable() const { return false; }
 
@@ -41,8 +41,8 @@ namespace Bousk
 						Multiplexer() = default;
 						~Multiplexer() = default;
 
-						void queue(std::vector<uint8_t>&& msgData);
-						size_t serialize(uint8_t* buffer, size_t buffersize, Datagram::ID);
+						void queue(std::vector<uint8>&& msgData);
+						uint16 serialize(uint8* buffer, uint16 buffersize, Datagram::ID);
 
 					private:
 						std::vector<Packet> mQueue;
@@ -55,8 +55,8 @@ namespace Bousk
 						Demultiplexer() = default;
 						~Demultiplexer() = default;
 
-						void onDataReceived(const uint8_t* data, size_t datasize);
-						std::vector<std::vector<uint8_t>> process();
+						void onDataReceived(const uint8* data, uint16 datasize);
+						std::vector<std::vector<uint8>> process();
 
 					private:
 						void onPacketReceived(const Packet* pckt);
